+++ unittests/worker_scripts/log_writer.py
@@ -31,15 +31,16 @@ def main():
         log_name = str(config[section]["log_name"])
         extension = str(config[section]["extension"])
         msg_types = str(config[section]["msg_types"])
-        max_days = float(config[section]["max_days"])
-        min_days = float(config[section]["min_days"])
-        max_mbytes = float(config[section]["max_mbytes"])
+        max_days = int(float(config[section]["max_days"]))
+        min_days = int(float(config[section]["min_days"]))
+        max_mbytes = int(float(config[section]["max_mbytes"]))
         backupCount = 5
         compression = ""
+        structured = False
 
         log_dir = "/tmp/%s" % log_name
 
-        logSupport.add_processlog_handler(
+        handler = logSupport.get_processlog_handler(
             log_name,
             log_dir,
             msg_types,
@@ -50,8 +51,12 @@ def main():
             backupCount=backupCount,
             compression=compression,
         )
+        if structured:
+            log = logSupport.get_structlog_logger(log_name)
+        else:
+            log = logSupport.get_logging_logger(log_name)
+        log.addHandler(handler)
 
-        log = logSupport.getLogger(log_name)
         log.info("%s\n" % create_random_string(length=2048))
 
         return 0